Quasiquoter creates data structure
The current code correctly expands a quasiquotation into a uriqq/8
term containing all the information (and variables) requested.  The
listing/0 call in t/examples.pl is to help me visually inspect that
it's happening correctly.

The next step is to use library(function_expansion) to convert uriqq/8
into an atom at runtime (when variable values are available).  The
current tests loop forever when they're run, so I'm guessing that
macro expansion is going wrong.